float votes

+++ assembl/models/votes.py
@@ -146,7 +146,7 @@ class LickertIdeaVote(AbstractIdeaVote):
 
     @value.setter
     def value(self, val):
-        val = int(val)
+        val = float(val)
         assert val <= self.lickert_range.maximum and \
             val >= self.lickert_range.minimum
         self.vote_value = val
